+++ src/thread_endpoint/send.rs
@@ -1,4 +1,4 @@
-use crossbeam_channel::Sender;
+use crossbeam_channel::{SendError, Sender};
 
 use crate::{
     pool_item::PoolItem, request_with_response::RequestWithResponse, sender_couplet::SenderCouplet,
@@ -12,13 +12,16 @@ where
     P: PoolItem,
 {
     /// This function send an asynchronous request to a thread pool
-    pub fn send<T>(&self, sender: &Sender<ThreadRequestResponse<P>>, request: T)
+    pub fn send<T>(
+        &self,
+        sender: &Sender<ThreadRequestResponse<P>>,
+        request: T,
+    ) -> Result<(), SendError<SenderCouplet<P>>>
     where
         T: RequestWithResponse<P>,
     {
         self.sender
             .send(SenderCouplet::<P>::new(sender.clone(), request))
-            .expect("The receiver thread to always be available");
     }
 }
 
@@ -50,7 +53,7 @@ mod tests {
         };
 
         // call send
-        target.send(&to_endpoint, echo_request.clone());
+        target.send(&to_endpoint, echo_request.clone()).unwrap();
 
         // get the message sent
         let sender_couplet = receiver_from_endpoint.recv().unwrap();